Unload on route, source data in App
This sort of resembles a service pattern where child elements request
data from the parent. The upside is we get to make child elements not
have to handle empty cases, they will unload on navigation, which
frees memory as well. The downside is that the App component has to
spy on signals from the child pages and provide data as necessary.